Scripting of Book Now from menu level
