+++ Feature/TopMenuPages.feature
@@ -5,3 +5,10 @@ Feature: Book Now Booking Form from Top Menu Home Page
     When The user clicks on the 'BOOK NOW' top menu
     Then User validates booking form is displayed
 
+    When The user clicks on the 'Rooms' top menu
+    Then User validates that Rooms page is open
+
+    When The user clicks on the 'Contact Us' top menu
+    Then User validates that Contact Us page is open
+
+
