+++ src/test/java/pages/MenuPage.java
@@ -6,5 +6,6 @@ public class MenuPage {
     public static By topMenu(String topMenu) {
         return By.xpath("//div/descendant::li[contains(@class,'nav-item')]/child::a[text()='" + topMenu + "']");
     }
+    public static By ContactUsHeading = By.xpath("//h2[text()='Contact Us']");
 
 }
